+++ src/module/data/message/ability-use.mjs
@@ -1,7 +1,13 @@
-import BaseMessageModel from "./base.mjs";
 
-/** @import AbilityModel from "../item/ability.mjs" */
-/** @import DrawSteelItem from "../../documents/item.mjs" */
+import BaseMessageModel from "./base.mjs";
+import DrawSteelActiveEffect from "../../documents/active-effect.mjs";
+
+/**
+ * @import { ActiveEffectData } from "@common/documents/_types.mjs";
+ * @import AbilityModel from "../item/ability.mjs";
+ * @import AppliedPowerRollEffect from "../pseudo-documents/power-roll-effects/applied-effect.mjs";
+ * @import DrawSteelItem from "../../documents/item.mjs"
+ */
 
 const fields = foundry.data.fields;
 
@@ -27,6 +33,27 @@ export default class AbilityUseModel extends BaseMessageModel {
 
   /* -------------------------------------------------- */
 
+  /**
+   * The displayed result tier.
+   * @type {1 | 2 | 3 | undefined}
+   */
+  get tier() {
+    // First roll is always the base roll, second is the result. Only one tier per message.
+    return this.parent.rolls[1]?.product;
+  }
+
+  /**
+   * The key of the tier.
+   * @type {"tier1" | "tier2" | "tier3" | null}
+   */
+  get tierKey() {
+    const tier = this.tier;
+    if (tier) return `tier${tier}`;
+    else return null;
+  }
+
+  /* -------------------------------------------------- */
+
   /** @inheritdoc */
   async alterMessageHTML(html) {
     await super.alterMessageHTML(html);
@@ -34,9 +61,7 @@ export default class AbilityUseModel extends BaseMessageModel {
     /** @type {DrawSteelItem & { system: AbilityModel}} */
     const item = await fromUuid(this.uuid);
 
-    // First roll is always the base roll, second is the result. Only one tier per message.
-    /** @type {1 | 2 | 3} */
-    const tier = this.parent.rolls[1]?.product;
+    const tierKey = this.tierKey;
 
     /** @type {HTMLDivElement} */
     const content = html.querySelector(".message-content");
@@ -46,7 +71,7 @@ export default class AbilityUseModel extends BaseMessageModel {
       let embed;
       if (item) {
         const embedConfig = {};
-        if (tier) embedConfig[`tier${tier}`] = true;
+        if (tierKey) embedConfig[tierKey] = true;
         embed = await item.toEmbed(embedConfig);
       }
       else {
@@ -57,10 +82,10 @@ export default class AbilityUseModel extends BaseMessageModel {
       // If it's a roll, the roll rendering will replace the message's stored content. Otherwise we need to do it.
       if (this.parent.isRoll) content.insertAdjacentElement("afterbegin", embed);
       else content.innerHTML = embed.outerHTML;
-    } else if (item && tier) {
+    } else if (item && tierKey) {
       content.insertAdjacentHTML("afterbegin", `<p class="powerResult"><strong>${
-        game.i18n.localize(`DRAW_STEEL.Roll.Power.Results.Tier${tier}`)
-      }: </strong>${item.system.powerRoll[`tier${tier}`].display}</p>`,
+        game.i18n.localize(`DRAW_STEEL.Roll.Power.Results.Tier${this.tier}`)
+      }: </strong>${item.system.power.effects.contents.map(effect => effect.toText(this.tier)).join("; ")}</p>`,
       );
     } else console.warn("Invalid configuration");
   }
@@ -70,14 +95,13 @@ export default class AbilityUseModel extends BaseMessageModel {
   /** @inheritdoc */
   async _constructFooterButtons() {
     const buttons = await super._constructFooterButtons();
+    const tier = this.tier;
     /** @type {DrawSteelItem & { system: AbilityModel}} */
     const item = await fromUuid(this.uuid);
+    if (!tier || !item) return buttons;
     for (const pre of item.system.power.effects) {
-      switch (pre.type) {
-        case "applied":
-          // TODO in next PR
-          break;
-      }
+      const newButtons = await pre.constructButtons(tier);
+      if (newButtons) buttons.push(...newButtons);
     }
     return buttons;
   }
@@ -87,5 +111,36 @@ export default class AbilityUseModel extends BaseMessageModel {
   /** @inheritdoc */
   addListeners(html) {
     super.addListeners(html);
+    /** @type {HTMLButtonElement[]} */
+    const effectButtons = html.querySelectorAll(".apply-effect");
+    for (const effectButton of effectButtons) effectButton.addEventListener("click", async (event) => {
+      /** @type {AppliedPowerRollEffect} */
+      const pre = await fromUuid(effectButton.dataset.uuid);
+      if (!pre) return void ui.notifications.error("DRAW_STEEL.Messages.AbilityUse.NoPRE", { localize: true });
+      const effectId = effectButton.dataset.effectId;
+      const config = pre.applied[this.tierKey].effects[effectId];
+
+      const noStack = !config.properties.has("stacking");
+
+      /** @type {DrawSteelActiveEffect} */
+      const tempEffect = effectButton.dataset.type === "custom" ?
+        pre.item.effects.get(effectId).clone({}, { keepId: noStack, addSource: true }) :
+        await DrawSteelActiveEffect.fromStatusEffect(effectId);
+
+      /** @type {ActiveEffectData} */
+      const updates = {
+        transfer: true,
+        origin: pre.uuid,
+        system: {},
+      };
+      if (config.end) updates.system.end = { type: config.end };
+      tempEffect.updateSource(updates);
+
+      for (const actor of ds.utils.tokensToActors()) {
+        // reusing the ID will block creation if it's already on the actor
+        // TODO: Update when https://github.com/foundryvtt/foundryvtt/issues/11898 is implemented
+        actor.createEmbeddedDocuments("ActiveEffect", [tempEffect.toObject()], { keepId: noStack });
+      }
+    });
   }
 }