+++ src/module/utils/construct-html-button.mjs
@@ -1,15 +1,16 @@
 /**
  * A helper method for constructing an HTML button based on given parameters.
  * @param {object} config Options forwarded to the button
- * @param {string} [config.label]
- * @param {Record<string, string>} [config.dataset={}]
- * @param {string[]} [config.classes=[]]
- * @param {string} [config.icon=""]
- * @param {"button" | "submit"} [config.type="button"]
- * @param {boolean} [config.disabled=false]
+ * @param {string} [config.label]                             A label for the button
+ * @param {Record<string, string>} [config.dataset={}]        Dataset info
+ * @param {string[]} [config.classes=[]]                      Classes to apply to the button
+ * @param {string} [config.icon=""]                           A FontAwesome icon for the button
+ * @param {string} [config.img=""]                            A URL image path. Takes priority over a FA icon.
+ * @param {HTMLButtonElement["type"]} [config.type="button"]  Button type
+ * @param {boolean} [config.disabled=false]                   Whether to disable the button
  * @returns {HTMLButtonElement}
  */
-export default function constructHTMLButton({ label = "", dataset = {}, classes = [], icon = "", type = "button", disabled = false }) {
+export default function constructHTMLButton({ label = "", dataset = {}, classes = [], icon = "", img = "", type = "button", disabled = false }) {
   const button = document.createElement("button");
   button.type = type;
 
@@ -17,9 +18,11 @@ export default function constructHTMLButton({ label = "", dataset = {}, classes 
     button.dataset[key] = value;
   }
   button.classList.add(...classes);
-  if (icon) icon = `<i class="${icon}"></i> `;
+  let image = "";
+  if (img) image = `<img src="${img}" alt="${label}">`;
+  else if (icon) image = `<i class="${icon}"></i> `;
   if (disabled) button.disabled = true;
-  button.innerHTML = `${icon}${label}`;
+  button.innerHTML = `${image}${label}`;
 
   return button;
 }